feat(phase3): analytics & view session lifecycle API
- Wire VoterController startView/completeView/viewHistory to ViewSessionResource
- Add withoutScopedBindings() to watch route (fixes Voter::campaigns() 500)
- Fix ViewSession::byStatus() to use mapWithKeys (enum key serialization)
- Fix scopePendingPayout to include Approved + Pending payment statuses
- Fix PoliticalViewService::assignView() Carbon addHours() type cast
- Add tests/Feature/Api/ViewSessionLifecycleTest.php (13 tests, 52 assertions)
  - Campaign availability, start view, fraud gate, view-goal 410
  - Progress heartbeat, qualifying/non-qualifying completion
  - Referral commission creation, view history, earnings summary
  - ViewSession::byStatus() analytics, scopePendingPayout scope
- Expand VoterApiTest watch endpoint assertion to include 429/410

108 tests passing, 0 failures

+++ app/Services/PoliticalViewService.php
@@ -42,7 +42,7 @@ class PoliticalViewService
                 'political_campaign_id' => $campaign->id,
                 'voter_id'              => $voter->id,
                 'status'                => ViewSessionStatus::Assigned,
-                'expires_at'            => Carbon::now()->addHours(config('u9itus.assignment_expiry_hours', 24)),
+                'expires_at'            => Carbon::now()->addHours((int) config('u9itus.assignment_expiry_hours', 24)),
                 'ip_address'            => $request->ip(),
                 'device_fingerprint'    => $request->header('X-Device-Fingerprint') ?? $request->input('device_fingerprint'),
                 'user_agent'            => $request->userAgent(),
